cleaned up code base - gabe
