--- ui/src/pages/Dashboard.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function Dashboard() {
-    return (
-        <div className="home">
-            <h1>Dashboard</h1>
-        </div>
-    )
-}
-
-export default Dashboard
