--- ui/src/pages/Reports.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function Reports() {
-    return (
-        <div className="reports">
-            <h1>Reports</h1>
-        </div>
-    )
-}
-
-export default Reports;
