--- ui/src/pages/Simulation.css
@@ -1,5 +1,0 @@
-.settings-container {
-    position: absolute;
-    top: 100px;
-    right: 0;
-}
